move behavior to traits
